requirements.txt
# Cricket Cover Drive Analysis - Requirements

# Web Application
streamlit>=1.28.0

# Computer Vision and Image Processing
opencv-python>=4.8.0

# Pose Estimation
mediapipe>=0.10.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
pandas>=2.0.0

# Video Download (for YouTube support)
yt-dlp>=2023.7.6

# Plotting
matplotlib>=3.7.0

# Progress Tracking
tqdm>=4.65.0

# Configuration Management
pyyaml>=6.0

# Environment Variables
python-dotenv>=1.0.0

# Type Hints Support
typing-extensions>=4.5.0
